fix] added not found error in announcemnts, from principals controller

+++ generic/app/controllers/announcements.rb
@@ -80,21 +80,24 @@ class Announcements < Application
        render :new
     end
   end
-  
+   
   def edit
-    @announcement = @current_school.announcements.find(params[:id])
+    @announcement = @current_school.announcements.find_by_id(params[:id])
+    raise NotFound unless @announcement
     @attachments = @current_school.attachments.find(:all, :conditions => ["attachable_id = ? and attachable_type =?", @announcement.id, "Announcement"])
     @allowed = 1 - @attachments.size
     render
   end
 
   def show
-    @announcement = @current_school.announcements.find(params[:id])
+    @announcement = @current_school.announcements.find_by_id(params[:id])
+    raise NotFound unless @announcement
     render
   end
 
   def update
-    @announcement = @current_school.announcements.find(params[:id])
+    @announcement = @current_school.announcements.find_by_id(params[:id])
+    raise NotFound unless @announcement
     @attachments = @current_school.attachments.find(:all, :conditions => ["attachable_id = ? and attachable_type =?", @announcement.id, "Announcement"])
     @allowed = 1 - @attachments.size
     i=0
@@ -155,7 +158,8 @@ class Announcements < Application
       @allowed = 1 - @attachments.size
       render :edit, :id => @announcement.id
     else
-      @announcement = @current_school.announcements.find(params[:id])
+      @announcement = @current_school.announcements.find_by_id(params[:id])
+      raise NotFound unless @announcement
       @page = @announcement.access_name
       Attachment.delete_all(['attachable_id = ?', @announcement.id])
       if @page == "Home Page"
